% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example1}
\alias{example1}
\title{Example data}
\format{
A data frame with 258 rows and 10 columns
}
\source{
Randomly generated by a mixed model with Treatment effect of -50
}
\usage{
example1
}
\description{
Example data
}
\keyword{datasets}
